fix(actions): move SMS_TEMPLATES const out of 'use server' module
Next.js 15 also forbids const object exports from 'use server' files,
not just sync functions. Production build fails at "Collecting page
data" with "A use server file can only export async functions, found
object." Move SMS_TEMPLATES (and its derived SmsTemplateId type) into
src/lib/sms/helpers.ts. Update SmsComposer.tsx import path.
